requirements.txt
requests>=2.0
flask>=3.0
openai>=1.0
python-dotenv>=1.0
beautifulsoup4>=4.12
trafilatura>=1.6
markupsafe>=2.1
pdfplumber>=0.10
python-docx>=1.0
striprtf>=0.0.26
playwright>=1.40
firecrawl-py>=1.6
